+++ CHANGELOG.md
@@ -7,6 +7,9 @@ but not always is possible (due the use of unstable features from Deno).
 Any BREAKING CHANGE between minor versions will be documented here in upper case.
 
 ## [1.18.5] - Unreleased
+### Added
+- JSON loader to esbuild [#473].
+
 ### Changed
 - Make Vento components synchronous.
 
@@ -2368,6 +2371,7 @@ The first version.
 [#467]: https://github.com/lumeland/lume/issues/467
 [#469]: https://github.com/lumeland/lume/issues/469
 [#470]: https://github.com/lumeland/lume/issues/470
+[#473]: https://github.com/lumeland/lume/issues/473
 
 [1.18.5]: https://github.com/lumeland/lume/compare/v1.18.4...HEAD
 [1.18.4]: https://github.com/lumeland/lume/compare/v1.18.3...v1.18.4